+++ src/tokio/conn.rs
@@ -482,7 +482,7 @@ impl Conn {
         params: P,
     ) -> Result<Option<Row>>
     where
-        Row: for<'buf> crate::raw::FromRawRow<'buf>,
+        Row: for<'buf> crate::raw::FromRow<'buf>,
         P: Params,
     {
         let result = self.exec_first_inner(stmt, params).await;
@@ -495,7 +495,7 @@ impl Conn {
         params: P,
     ) -> Result<Option<Row>>
     where
-        Row: for<'buf> crate::raw::FromRawRow<'buf>,
+        Row: for<'buf> crate::raw::FromRow<'buf>,
         P: Params,
     {
         write_execute(self.buffer_set.new_write_buffer(), stmt.id(), params)?;
@@ -521,7 +521,7 @@ impl Conn {
         params: P,
     ) -> Result<Vec<Row>>
     where
-        Row: for<'buf> crate::raw::FromRawRow<'buf>,
+        Row: for<'buf> crate::raw::FromRow<'buf>,
         P: Params,
     {
         let mut handler = crate::handler::CollectHandler::<Row>::default();
@@ -539,7 +539,7 @@ impl Conn {
         f: F,
     ) -> Result<()>
     where
-        Row: for<'buf> crate::raw::FromRawRow<'buf>,
+        Row: for<'buf> crate::raw::FromRow<'buf>,
         P: Params,
         F: FnMut(Row) -> Result<()>,
     {